Added 'play', 'quit', 'volume', 'quality', and 'fullscreen' features
The main menu script gives functionality to the main menu allowing for the 'play game' and 'quit game' features. The 'play game' feature changes to the second scene which starts the game. The 'quit game' feature exits out of the game. The settings menu script gives functionality to the settings menu allowing for the 'set volume', set 'quality', and 'set fullscreen' features. The 'set volume' feature changes the volume to the value given to the slider. The 'set quality' feature changes the quality of the game which is changed through the drop down. The 'fullscreen' feature toggles between fullscreen based on the value of the toggle box.