Added pause menu script
This script allows the pause menu scene to pop up when the 'ESC' key is pressed (while the game is running). This script contains the 'resume' and 'pause' method features. When the game is paused the game time will stop/freeze. The console will be messaged when the game is paused or resumed.