Added settings menu script
The settings menu script gives functionality to the settings menu allowing for the 'set volume', set 'quality', and 'set fullscreen' features. The 'set volume' feature sets the volume to the value given to the slider (which is default 0). The 'set quality' feature allows the user to change the quality of the game (default is HIGH). The 'set fullscreen' feature allows the user to toggle between fullscreen and windowed based on the value given to the toggle.