Added inventory script
This script allows the inventory to pop up when the 'TAB' key is pressed (while the game is running). This script contains the 'Open Inventory' and 'Close Inventory' method features. When the inventory is open the game time will stop/freeze. The console will be messaged when the inventory is opened/closed.