Scripts for New Inventory - see longer descripton
Inventory - like the previous inventory script, but most of it is commented out.  The "press TAB" functionality was kept the same and works.  PutItemInInventory, which puts an item in the inventory, has not yet been tested
Inventory Slots - the slots in the inventory.  This is where an item would actually go once picked up.  Not yet tested.