updated prefabs - see longer description
The Game Manager and Inventory prefabs have not been saving properly in git.  If they throw errors in the console:
For the Game Manager, go into the prefab, put the dialogSystem script in its script place, and put the text box and the text area into their respective places from the childed dialog box.
For the Inventory, go into the ItemSlot prefab.  InventoryItems should be attached to Icon and ItemSlots should be attached to the ItemSlot itself. 
Those should clear up any errors.